+++ app/src/main/java/com/boostcamp/and03/ui/screen/booklist/BooklistViewModel.kt
@@ -2,7 +2,7 @@ package com.boostcamp.and03.ui.screen.booklist
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.boostcamp.and03.data.repository.book_storage.BookStorageRepository
+import com.boostcamp.and03.data.repository.bookstorage.BookStorageRepository
 import com.boostcamp.and03.ui.screen.booklist.model.toUiModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.collections.immutable.toImmutableList
